Reorganized the classes in respective folders!
