+++ Octokit.Tests.Integration/Octokit.Tests.Integration.csproj
@@ -91,7 +91,8 @@
     <Compile Include="Clients\UserEmailsClientTests.cs" />
     <Compile Include="Clients\FollowersClientTests.cs" />
     <Compile Include="Clients\UserKeysClientTests.cs" />
-    <Compile Include="GithubClientExtensions.cs" />
+    <Compile Include="Helpers\GithubClientExtensions.cs" />
+    <Compile Include="Helpers\DisposableRepository.cs" />
     <Compile Include="Helpers\PaidAccountTestAttribute.cs" />
     <Compile Include="Helpers\RepositorySetupHelper.cs" />
     <Compile Include="HttpClientAdapterTests.cs" />
